feat : Added print_img
